refac: move HostMonitor to a new module
