+++ tests/system/test_active_response_log_format/test_active_response_log_format.py
@@ -45,7 +45,7 @@ import time
 import pytest
 from random import randint
 from wazuh_testing.tools import WAZUH_LOGS_PATH
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager, clean_environment
