+++ tests/system/test_cluster/test_agent_enrollment/test_agent_enrollment.py
@@ -6,7 +6,7 @@ import os
 
 import pytest
 from wazuh_testing.tools import WAZUH_PATH, WAZUH_LOGS_PATH
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager
