+++ tests/system/test_cluster/test_agent_files_deletion/test_agent_files_deletion.py
@@ -8,7 +8,7 @@ from time import time, sleep
 
 import pytest
 from wazuh_testing.tools import WAZUH_PATH, WAZUH_LOGS_PATH
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager
