+++ tests/system/test_cluster/test_agent_info_sync/test_agent_info_sync.py
@@ -9,7 +9,7 @@ from time import sleep, time
 
 import pytest
 from wazuh_testing.tools import WAZUH_PATH, WAZUH_LOGS_PATH
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager
