+++ tests/system/test_cluster/test_agent_key_polling/test_agent_key_polling.py
@@ -6,7 +6,7 @@ import os
 
 import pytest
 from wazuh_testing.tools import WAZUH_LOGS_PATH
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager
