+++ tests/system/test_cluster/test_integrity_sync/test_integrity_sync.py
@@ -12,7 +12,7 @@ import pytest
 import yaml
 
 from wazuh_testing.tools import WAZUH_PATH, PYTHON_PATH, WAZUH_LOGS_PATH
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager
