+++ tests/system/test_fim/test_files/test_files_cud.py
@@ -43,7 +43,7 @@ tags:
 import os
 
 import pytest
-from wazuh_testing.tools.monitoring import HostMonitor
+from wazuh_testing.tools.system_monitoring import HostMonitor
 from wazuh_testing.tools.system import HostManager, clean_environment
 from wazuh_testing.tools import WAZUH_LOGS_PATH
 from wazuh_testing.fim import create_folder_file, wait_for_fim_scan_end
